backend/requirements/base.txt
# Core runtime dependencies (minimal deterministic set)
APScheduler==3.11.1
Jinja2==3.1.6
MarkupSafe==3.0.3
PyJWT==2.10.1
PyYAML==6.0.3
aiohttp-retry==2.9.1
aiohttp==3.13.2
annotated-types==0.7.0
anyio==4.12.0
bcrypt==4.1.3
boto3==1.42.5
botocore==1.42.5
click==8.3.1
cryptography==46.0.3
dnspython==2.8.0
elevenlabs==2.31.0
email-validator==2.3.0
fastapi==0.110.1
fastuuid==0.14.0
httpx==0.28.1
motor==3.3.1
openai==1.99.9
passlib==1.7.4
pydantic==2.12.5
pymongo==4.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
requests==2.32.5
starlette==0.37.2
stripe==14.0.1
tenacity==9.1.2
twilio==9.8.8
typer==0.20.0
typing_extensions==4.15.0
uvicorn==0.25.0
watchfiles==1.1.1
websockets==15.0.1

# Core platform additions
redis==5.0.7
celery[redis]==5.4.0
slowapi==0.1.9
pyotp==2.9.0
